rename BarChartWithLabels folder to BarCharts
