+++ src/components/charts/new_revocations/BarCharts/BarChartWithLabels.js
