+++ src/components/charts/new_revocations/BarCharts/HorizontalBarChartWithLabels.js
