+++ src/components/charts/new_revocations/BarCharts/helpers.js
