+++ src/components/charts/new_revocations/BarCharts/index.js
